+++ lib/unrouter.dart
@@ -38,7 +38,7 @@ export 'src/guard.dart';
 export 'src/inlet.dart';
 export 'src/extensions.dart';
 export 'src/navigation.dart';
+export 'src/route_state.dart';
 export 'src/route_matcher.dart' show MatchedRoute;
 export 'src/router.dart';
-export 'src/router_state.dart';
 export 'src/url_strategy.dart';